fix to build on linux
git-svn-id: http://svn.haeleth.net/onscr/trunk@232 9001f46f-672a-0410-b4fc-d4ff93ad6438

+++ ONScripterLabel_sound.cpp
@@ -762,7 +762,7 @@ static size_t oc_read_func(void *ptr, size_t size, size_t nmemb, void *datasourc
     OVInfo *ogg_vorbis_info = (OVInfo*)datasource;
 
     size_t len = size*nmemb;
-    if (ogg_vorbis_info->pos+len > ogg_vorbis_info->length) 
+    if ((size_t)ogg_vorbis_info->pos+len > (size_t)ogg_vorbis_info->length) 
         len = (size_t)(ogg_vorbis_info->length - ogg_vorbis_info->pos);
     memcpy(ptr, ogg_vorbis_info->buf+ogg_vorbis_info->pos, len);
     ogg_vorbis_info->pos += len;
